+++ data/restclient/src/main/java/com/freshdigitable/udonroad2/data/restclient/TweetApiClient.kt
@@ -28,29 +28,3 @@ class TweetApiClient @Inject constructor(
         twitter.retweetStatus(id.value).toEntity()
     }
 }
-
-class AppTwitterException constructor(exception: TwitterException) : Exception(exception) {
-    val statusCode: Int = exception.statusCode
-    val errorCode: Int = exception.errorCode
-    val errorType: ErrorType? = findByCode(statusCode, errorCode)
-    val exceptionCode: String = exception.exceptionCode
-
-    companion object {
-        private const val STATUS_FORBIDDEN: Int = 403
-
-        private fun findByCode(statusCode: Int, errorCode: Int): ErrorType? {
-            return ErrorType.values().first {
-                it.statusCode == statusCode && it.errorCode == errorCode
-            }
-        }
-    }
-
-    enum class ErrorType(
-        val statusCode: Int,
-        val errorCode: Int
-    ) {
-        // https://developer.twitter.com/ja/docs/basics/response-codes
-        ALREADY_FAVORITED(STATUS_FORBIDDEN, 139),
-        ALREADY_RETWEETED(STATUS_FORBIDDEN, 327),
-    }
-}
